draft array converter
