+++ flopy4/mf6/component.py
@@ -1,0 +1,10 @@
+from abc import ABC
+
+COMPONENTS = {}
+"""MF6 component registry."""
+
+
+class Component(ABC):
+    @classmethod
+    def __attrs_init_subclass__(cls):
+        COMPONENTS[cls.__name__.lower()] = cls
