+++ flopy4/mf6/constants.py
@@ -1,0 +1,4 @@
+import numpy as np
+
+FILL_DEFAULT = np.nan
+FILL_DNODATA = 1e30
