+++ flopy4/mf6/gwf/ic.py
@@ -1,8 +1,10 @@
 import numpy as np
+from attrs import Converter
 from numpy.typing import NDArray
 from xattree import array, field, xattree
 
 from flopy4.mf6 import Package
+from flopy4.mf6.converters import convert_array
 
 
 @xattree
@@ -11,6 +13,7 @@ class Ic(Package):
         dims=("node",),
         default=1.0,
         metadata={"block": "packagedata"},
+        converter=Converter(convert_array, takes_self=True, takes_field=True),
     )
     export_array_ascii: bool = field(
         default=False, metadata={"block": "options"}
