+++ flopy4/mf6/model.py
@@ -1,0 +1,8 @@
+from attrs import define
+
+from flopy4.mf6.component import Component
+
+
+@define
+class Model(Component):
+    pass
