+++ flopy4/mf6/simulation.py
@@ -1,0 +1,15 @@
+from xattree import field, xattree
+
+from flopy4.mf6.component import Component
+from flopy4.mf6.exchange import Exchange
+from flopy4.mf6.model import Model
+from flopy4.mf6.solution import Solution
+from flopy4.mf6.tdis import Tdis
+
+
+@xattree
+class Simulation(Component):
+    models: dict[str, Model] = field()
+    exchanges: dict[str, Exchange] = field()
+    solutions: dict[str, Solution] = field()
+    tdis: Tdis = field()
