+++ flopy4/mf6/solution.py
@@ -1,0 +1,8 @@
+from attrs import define
+
+from flopy4.mf6.package import Package
+
+
+@define
+class Solution(Package):
+    pass
